No simd
