--- SIMD/AlphaBlending/blend_images.cpp
@@ -1,3 +1,0 @@
-#include "blend_images.hpp"
-
-
