--- SIMD/AlphaBlending/blend_images.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-#include "RGBA.hpp"
-
-struct img
-  {
-  RGBA* pixel_arr;
-
-  unsigned width;
-  unsigned height;
-  };
-
-
