+++ SIMD/AlphaBlending/include/blend_images.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "Image.hpp"
+
+int Blend (Image* front, Image* background, unsigned x_coord, unsigned y_coord);
